🚀 fix(ci): add --allow-dirty flag to bump-my-version in publish workflow
The publish workflow was failing because bump-my-version chokes on dirty
working directories. Since we intentionally modify pyproject.toml to change
the package name from 'gac' to 'gac-dev' before publishing, we need to tell
bump-my-version to chill out with the --allow-dirty flag.

Now our CI can bump versions like the absolute unit it was meant to be! 💪